Fix comment: Move back MultiSet; define T? key.
