+++ src/libraries/System.Collections.Immutable/src/System.Collections.Immutable.csproj
@@ -94,7 +94,6 @@ System.Collections.Immutable.ImmutableStack&lt;T&gt;</PackageDescription>
     <Compile Include="System\Collections\Immutable\ImmutableStack_1.cs" />
     <Compile Include="System\Collections\Immutable\ImmutableStack_1.Enumerator.cs" />
     <Compile Include="System\Collections\Immutable\KeysOrValuesCollectionAccessor.cs" />
-    <Compile Include="System\Collections\Immutable\MultiSet.cs" />
     <Compile Include="System\Collections\Immutable\RefAsValueType.cs" />
     <Compile Include="System\Collections\Immutable\SecureObjectPool.cs" />
     <Compile Include="System\Collections\Immutable\SortedInt32KeyNode.cs" />
