+++ src/libraries/System.Collections.Immutable/src/System/Collections/Immutable/ImmutableArray_1.cs
@@ -6,6 +6,7 @@ using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
 using System.Globalization;
 using System.Linq;
+using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using System.Runtime.Versioning;
 
@@ -576,6 +577,7 @@ namespace System.Collections.Immutable
             self.ThrowNullRefIfNotInitialized();
             Requires.NotNull(items, nameof(items));
 
+            // Avoid to allocate for true enumerable
             ICollection<int>? indicesToRemove = items.TryGetCount(out _) ? self.FindOrderedIndicesToRemoveByMultiSet(items, equalityComparer) : self.FindOrderedIndicesToRemove(items, equalityComparer);
             return indicesToRemove == null ? self : self.RemoveAtRange(indicesToRemove);
         }
@@ -1385,7 +1387,7 @@ namespace System.Collections.Immutable
 
         private ICollection<int>? FindOrderedIndicesToRemoveByMultiSet(IEnumerable<T> items, IEqualityComparer<T>? equalityComparer)
         {
-            var multiSet = new MultiSet<T>(equalityComparer);
+            var multiSet = new MultiSet(equalityComparer);
             foreach (T item in items)
             {
                 multiSet.Add(item);
@@ -1396,7 +1398,7 @@ namespace System.Collections.Immutable
 
         private ICollection<int>? FindOrderedIndicesToRemoveByMultiSet(ReadOnlySpan<T> items, IEqualityComparer<T>? equalityComparer)
         {
-            var multiSet = new MultiSet<T>(equalityComparer);
+            var multiSet = new MultiSet(equalityComparer);
             foreach (ref readonly T item in items)
             {
                 multiSet.Add(item);
@@ -1405,7 +1407,7 @@ namespace System.Collections.Immutable
             return GetOrderedIndicesToRemoveFor(multiSet);
         }
 
-        private ICollection<int>? GetOrderedIndicesToRemoveFor(MultiSet<T> multiSet)
+        private ICollection<int>? GetOrderedIndicesToRemoveFor(MultiSet multiSet)
         {
             List<int>? indicesToRemove = null;
             for (int i = 0; i < array!.Length; i++)
@@ -1432,6 +1434,81 @@ namespace System.Collections.Immutable
             }
 
             return indicesToRemove;
+        }
+
+        private readonly struct MultiSet
+        {
+            private readonly Dictionary<NullableKeyWrapper, int> _dictionary;
+
+            public MultiSet(IEqualityComparer<T>? equalityComparer)
+            {
+                _dictionary = new Dictionary<NullableKeyWrapper, int>(new NullableKeyWrapperEqualityComparer(equalityComparer ?? EqualityComparer<T>.Default));
+            }
+
+            public void Add(T? item)
+            {
+#if NET6_0_OR_GREATER
+                ref int count = ref CollectionsMarshal.GetValueRefOrAddDefault(_dictionary, item, out _);
+                count++;
+#else
+            _dictionary[item] = _dictionary.TryGetValue(item, out int count) ? count + 1 : 1;
+#endif
+            }
+
+            public bool TryRemove(T? item)
+            {
+#if NET6_0_OR_GREATER
+                ref int count = ref CollectionsMarshal.GetValueRefOrNullRef(_dictionary, item);
+                if (Unsafe.IsNullRef(ref count) || count == 0)
+#else
+            if (!_dictionary.TryGetValue(item, out int count) || count == 0)
+#endif
+                {
+                    return false;
+                }
+
+#if NET6_0_OR_GREATER
+                count--;
+#else
+            _dictionary[item] = count - 1;
+#endif
+                return true;
+            }
+
+            private readonly struct NullableKeyWrapper
+            {
+                public readonly T? Key;
+
+                public static implicit operator NullableKeyWrapper(T? key)
+                {
+                    return new NullableKeyWrapper(key);
+                }
+
+                private NullableKeyWrapper(T? key)
+                {
+                    Key = key;
+                }
+            }
+
+            private class NullableKeyWrapperEqualityComparer : IEqualityComparer<NullableKeyWrapper>
+            {
+                private readonly IEqualityComparer<T> _keyComparer;
+
+                public NullableKeyWrapperEqualityComparer(IEqualityComparer<T> keyComparer)
+                {
+                    _keyComparer = keyComparer;
+                }
+
+                public int GetHashCode(NullableKeyWrapper obj)
+                {
+                    return obj.Key == null ? 0 : _keyComparer.GetHashCode(obj.Key);
+                }
+
+                public bool Equals(NullableKeyWrapper x, NullableKeyWrapper y)
+                {
+                    return _keyComparer.Equals(x.Key, y.Key);
+                }
+            }
         }
     }
 }
